Fix duplicate getPriceMod
